Add classes into packages
