+++ src/main/java/arthur/Storage.java
@@ -1,3 +1,7 @@
+package arthur;
+
+import arthur.task.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
