+++ src/main/java/arthur/Ui.java
@@ -1,3 +1,5 @@
+package arthur;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
