+++ src/main/java/arthur/exceptions/EmptyDescriptionException.java
@@ -1,5 +1,7 @@
+package arthur.exceptions;
+
 /**
- * Custom exception for Arthur.
+ * Custom exception for arthur.Arthur.
  * Thrown when supported instruction doesn't have description.
  */
 public class EmptyDescriptionException extends Exception {
