+++ src/main/java/arthur/task/Task.java
@@ -1,8 +1,10 @@
+package arthur.task;
+
 /**
  * A class that creates task objects
  */
 public class Task {
-    private final String description;     // Task description
+    private final String description;     // Task.Task description
     private boolean isDone;
 
     public Task(String e) {
