+++ text-ui-test/EXPECTED.TXT
@@ -40,7 +40,7 @@ __________________________________________________
 The description of event cannot be empty
 __________________________________________________
 __________________________________________________
-Added a new Todo task: 
+Added a new Todo task:
 [T][ ] test
 You have 1 tasks in list at the moment.
 __________________________________________________
@@ -53,7 +53,7 @@ __________________________________________________
 Please add the deadline date
 __________________________________________________
 __________________________________________________
-Added a new Deadline task: 
+Added a new Deadline task:
 [D][ ] test2 (By: mon)
 You have 2 tasks in list at the moment.
 __________________________________________________
@@ -67,7 +67,7 @@ __________________________________________________
 Invalid task number. Please try again
 __________________________________________________
 __________________________________________________
-Good job! Task Completed 
+Good job! Task Completed
 [D][X] test2 (By: mon)
 __________________________________________________
 __________________________________________________
@@ -77,7 +77,7 @@ __________________________________________________
 Please add the event Date and/or Time
 __________________________________________________
 __________________________________________________
-Added a new Event task: 
+Added a new Event task:
 [E][ ] test3 (At: tues 5pm)
 You have 3 tasks in list at the moment.
 __________________________________________________
